Changes to Drag n Drop api
capDraggable() now takes a capability rather than a function, and doesn't
automagically grant a capability when a drag starts.

capDroppable() doesn't do a cap invocation on the dropped cap anymore, and
hands it directly to the drop target.

Content negotiation with highlighting still works, and the library avoids
passing caps with mismatched RCs through.  However, the drop target and drag
starter are responsible for any extra-info content negotiation --- e.g. the
extra generate step of buzzer, which emote.js got an extra invoke to handle.