extend mock'd web server to support XHR API
This gets the bulk of the unit tests for caps.js working again. caps.js stopped
using jQuery.ajax() in favor of XmlHttpRequest. This change extends the mock'd
web server in the unit tests to support the XHR API.

Small other change - CapServer.grant(<url>) changed semantics -- it now *wraps*
the BCAP, so one of the unit tests was wrong.