add interlock on the belay.onportready functions being called
Under Firefox the frame was sending the outpost (initial packet of caps) to the
page BEFORE the page had 'woken up', noticed the belayPort was set and
established the CapTunnel. This caused the outpost to be dumped on the floor.

In this CL, the frame is not notified of the page features (startId, location)
until after it is certain that the belay port has been dealt with by the page.

The change is that the actionPort post from belay-port.js (in the page) to the
iframe is delayed until after the confirmation callback is called. This
confirmation callback now becomes part of the API contract between belay-port
and the page.