remove active close notification
The main aim of this CL is to remove the flaky system whereby the station kept
track of which instances were open and which were closed via an active close
notification sent by the instance before it unloaded. This was flaky at best, and
failed Firefox entirely.

The sub-parts of this CL are:

- remove notifyClose / closeInstHandler logic
- remove instance.opened logic
- offer stations an isRunning cap to test if an instance is running
- plumb "navigation" structure (navigate & close caps) to buildActivateCap...
    ...so the activate cap can be told to close, rather than activate
- merge launchPageInstance and launchInstance, and remove old gadget cruft (finally!)
- check for already running in launchInstance (based on isRunning)...
   ...not in readyCap (based on .opened), and if running, tell activate cap to close
- filter suggestions by isRunning, not by .opened