support a page activating an instance on itself
Enables a third kind of launch, where we store all the launch info into the
start instruction. This allows a page to launch without having to call back to
the launching belay-frame to get the info. The downside is that we store
sensitive data into local storage with less assurance that it will be picked up
and removed.

This kind of launch is used when the user logs into station through some
credential: The station server's index page is replaced by the newly launched