auto-relaunch of station, and sign-out button
Station now remembers in localStorage the last station launched. Then when
visiting the index.html page, it will auto launch into the user's station if
found.

The station now has a "Sign out" link (upper right), that will clear this link,
and return to the index page. If there are no identities associated with this
station that can be used to log in, there is a (janky) confirmation dialog. Real
users will never see this, as you need to give some identity to create the
station with in the first place.

Lastly, some small simplification of the interlock when logging in  / creating
via open id: we can close the pop-up sooner now that we have self-activate
working.