check/enforce origin on messages from/to belay-frame
This CL is half the work of ensuring that postMessages aimed at windows and
iframes have the proper origin checks.

This CL changes belay-port.js so that it sets and checks the bely-belay origin
on port messages to/from the belay iframe. Both comm styles had to be adjusted.

The interface to the comms was altered slightly to so that all the window
postMessage handling could be confined within the comms implementations. In
particular, when the comms is to receive the first message (as it does in
belay-port) then the comms adds the handler and processes the first event. (The
case where the comms is to send the first message hasn't changed, and was
already managed by the comms.)

This change was propagated to belay-frame's version, but doesn't affect it,
since belay-frame sends the first message.

Issue 5522051