refactor of belay.py to support threaded AppEngine
This is, alas, a large CL, thought 90% of the changes are mechanical. The basic
aim is that many of the operations in belay.py were just free functions
operating on global state. While convenient, this isn't practical in a threaded,
webapp2 environment. (Thought AppEngine tries valiantly!)

* Created a CapServer object, and one is added to each instantiated BcapHandler
  - grant, regrant, data_pre/post_process, restore, are now methods of
CapServer
  - this resulted in many simple changes to call these via self.cap_server

* server_url() is now a method of BcapRequest
  - this resulted in many simple changes to call server_url via self

* Since grant, regrant, and server_url are no longer global, occasionally some
re-routing of logic was required to ensure we had a handler or cap_server in
scope.

* Removed outdated, unused "tools" logic in station. (Hold over from embedded
tools in early demos.)

* Reworked unit test of belay.py significantly for webapp2