refactor all js files to use require.js to load
This large CL is the first of a several phased sequence changing the way the .js
code is partitioned, and how it is loaded. The aim is to make it all
modularlized and use require.js to load it all.

This particular CL makes changes the basic .js libraries, and minimal changes to
the unit tests to get them running. It does not make the unit tests themselves
use require.js, nor do any of the belay apps (all of which are now broken).