Second change set of conversion to new API
This change converts the apps and tests to use the new method for including the
belay client, and for starting belay. In some cases, legacy calls have been used
- those can come out in a subsequent CL.

Most changes are simple. Two, belay/index.js and station/your-stuff.js should be
reviewed in an diff program that can handle changes